Add uTmax user files
Add user_files directory that contais example calibration (cal.txt),
valve tube database (data.csv) and plot output files.

However, these files may be out dated because the live files need
to be in /home/$USER/uTmax_files because this is a hard-coded path
in the program. The program will fail to run if the path does not
exist.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>