mainwindow: Reduce tube selector to font 8
It became hard to read the long entries in the tube selector pull
down list because the centre text was replaced with ... as the font
was too big.

Therefore, reduce the font from 9 to 8 so the whole text can be read.

Signed-off-by: Dean <skullandbones99@gmail.com>
(cherry picked from commit b790c62eaeab5e49a0348ba34710a3840893ffb1)