Bug: Fix failure to check home directory path exists
The mainwindow constructor defines the path to where the tube data
file and the calibration file are located but fails to check the
path exists.

Without the check, an attempt to write a new calibration file
silently fails.

Therefore, add a check to ensure the path to the data files exists
otherwise exit with an error message.

In addition, add some debug.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>