mainwindow: ReadCalibration() start with default calibration
It was noted that the initial default calibration file was
using a slightly different format to SaveCalFile().
The default wrote in float format "1.0" but SaveCalFile()
wrote "1". Also the calData.Vg40Val entry was duplicated in the
default write case.

Therefore, simplify ReadCalibration() by:

a) start with the default calibration information
b) if no calibration file exists use SaveCalFile() to save it
c) read the existing calibration file

This solution ensures there is a single function to read and
a single function to write the calibration information so
removes duplicate code and makes the code less error prone.

In addition, remove the serial port check code from SaveCalFile()
because comport will have the name of the user selected serial port
or the default first serial port name or "none" if no serial ports
are available. This means the check was unnecessary.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>
(cherry picked from commit 533ba85075400ce38094b32bb06f020c1c10289c)
Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>