Bug: Do not update the calibration file on every start
It is unnecessary to update the calibration file every time
the program starts because the calibration file has been read
before opening the serial port. Therefore, no information changed
so rewriting the file with the same information is redundant.

The solution is to add a flag updateCalFile parameter to
OpenComPort() to control whether the calibration file is
written to when opening the serial port.

Modify SerialPortDiscovery() to not save the calibration file.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>
(cherry picked from commit b58cd9a55cacc92b2e03313ad76b6021bf30e31a)
Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>