debug_dialog: Optimise opening serial port
Avoid unnecessarily opening the current same serial port and
consequently updating the calibration file with no changes.

This is done by adding a flag currentSerialPort to
UpdateComSel() which tracks whether the current serial port
has been selected.

If the current serial port is not found in the list, then
automatically select the first available serial port.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>