mainwindow: Add CloseComPort
Create a separate function CloseComPort() to allow the
serial port to be closed independently of open such as
when the USB to serial adaptor disappears or when the mainwindow
destructor is called.

Make the lifefime of the portInUse object match the port open
status. Therefore, portInUse is NULL when the serial port is not
in the open state.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>