mainwindow: Implement new serial protocol state machine
Add SendCommand() to implement a state machine to handle
the Data Protocol of the uTracer.

The state machine manages the following:

1. Manages TX and RX states transitions.

2. Transmits single characters of the command message and checks that
   the character is echoed. Therefore, a communications failure can
   be detected such as the uTracer not being present.

3. Handles switches between sending a command message and, if expected,
   receiving a message. This is half-duplex; a command and response
   cannot occur at the same time.

4. Uses a structure to hold the messages and states. This allows for
   potential flexibility such as dynamic allocation of the structures.

This solution is an improvement on the original design because this
design is more concise. Also, the design follows the design of the protocol
more closely such as enforcing half-duplex transmission / reception.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>
(cherry picked from commit 0c600bcfddbc16c37e2fa042ff5ead645578d2dd)
Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>