mainwindow: Modify RxPkt() to use new protocol state machine
Modify RxPkt() to use the state outputs of sendCommand()
instead of using the original solution.

The original protocol timeout mechanism was retained.

Note that the error case RXINVALID is no longer output although
might be reused in future for handling echo failure.

The statment
timeout = PING_TIMEOUT;
was removed from the heater control code because the timer was
expiring when no message had been sent.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>