mainwindow: Integrating user "start" to the state machine
Eliminate startSweep by adding user request "start" to the state
machine.

Depending on the current state of the state machine, the user
"start" request will be actioned to perform:

1. start the test - heater voltage starts to ramp up
2. jump to full heater voltage
3. start the sweep without fully waiting for the 1 minute heat-up
4. repeat the sweep test - heater remains on

Therefore, pressing the "start" button during the various phases
of the test run will jump to the next part of the test run.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>