mainwindow: Replace "heat" with "timeIt" in RxData()
It seems "heat" was set to 0 in multiple places but in fact it
was redundant because state Heating_wait_adc set "heat" to 1
at the start of the heater ramp-up.

Therefore, remove all heat = 0

In addition, reduce the scope of "heat" to RxData() by moving
the declaration of "heat" from the class into RxData().

There is no need to have multiple time variables in the state
machine so replace "heat" with "timeIt".

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>
(cherry picked from commit f99fb9026d1bafae8c2932b144b49af99f83c377)
Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>