mainwindow: Replace TIMER_SET with STATE_PERIOD
Improve the meaning of the #define TIMER_SET by renaming it to
STATE_PERIOD to indicate the tick time in ms.

In addition, add #define TICKS_PER_SEC to indicate the number
of increments per second to make it clearer what the time
measurements are in terms of ticks per second.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>