mainwindow: Remove unnecessary state "wait_stop"
Because the "HeatOff" state now checks the RXCODE, it makes
the "wait_stop" unnecessary so remove the "wait_stop" state.

In other words, the "HeatOff" state will not perform any action
until the current communications has completed and that was
the purpose of the "wait_stop" state.

In addition, add return statements in the switch statement
after starting to send the filament off command. This is needed
to allow the command to be fully sent before the next tick
of the state machine occurs.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>