Add updated calibration file
Ran the calibration procedure and updated the calibration file.

It was noted that some calibration values change slightly when
saving and reading back the calibration file. Suspect this is
due to using floating point calculations.

The Vg1 and Vg4 values are now set outside the original dynamic
range so needs the commit to expand the range of the calibration
sliders from the original range of 0.8 to 1.2 to the new range
of 0.7 to 1.3.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>