Improve discharging on "stop" and loss of comms
When a communications failure is detected, enter CommsFail
state and wait for a user request such as "stop" or "start".
Then send an End Measurement command in an attempt to
discharge the uTracer.

Use state WaitGoodComms to wait for successful communications
to occur or the communications timeout occurs.

Modify the HeatOff state to always send an End Measurement
command so that "stop" should always ensure the uTracer
is discharged except when communications fail.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>