BUG: Fix Vg1 calibration slider scaling
It seems that Cal_Dialog::on_Vg1Sli_valueChanged()
doubled the value of the offset when the slider
was moved. This caused the max displayed value to
change from 1.2 to 1.4 and the min displayed value
to change from 0.8 to 0.6. Therefore, the Vg1 reading
in the dialog is shown incorrectly after the slider
is moved.

The solution was to remove the x2 in
Cal_Dialog::on_Vg1Sli_valueChanged() which makes it
consistent with the other sliders.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>