Increase calibration slider range
My uTracer needed:
Vg1 to be 1.2366 and
Vg4 to be 1.2258

but the sliders were restricted to a maximum of 1.2 which
prevented the needed values from being set.

The solution was to modify the range from 0.2 to 0.3 and
to increase the number of steps from 2000 to 3000. This
causes the maximum slider value to be 1.3 and the minimum
slider value to be 0.7.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>