BUG: Fix gm1Del typo in saving data.csv
There is a typo in the first line written to the data.csv tube
data file. The 2nd "gm1Del" should be "gm2Del" and this commit
fixes the typo.

Note that when the first line of the data.csv tube data file
is read, the line is discarded. This means the typo has no
adverse side effects and the typo can be fixed with no impact.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>