BUG: Reduce quantisation noise in plotting Va
Va is an output and is not measured. The sweep Va value is
converted to an ADC value with some compensation for circuit
voltage drops. This is a "many to one" function because
the ADC can only represent specific voltages nearest to the
desired Va value. The Va step value may cause the same ADC
value to be used for consecutive desired Va sweep values.

However the current solution creates a "real" Va by reverse
calculating from the Va ADC value including some circuit voltage
drop compensations.

By plotting the "real" Va values, quantisation noise is introduced
into the Va values causing a 2 dimensional error box for each point
plotted.

But this is the wrong approach because when plotting Ia against
Va the error should only be represented in the Ia axis to have
all the error in a single dimension such as in a histogram.

Therefore, modify the plot to record the desired Va sweep value
against the measured Ia value. This means that when 2 Va sweep
values use the same ADC value, the measured Ia value should
be similar leading to the plot showing a short horizontal line.
This artifact is more desirable than the 2 dimensional error
box which can cause the gaps between plotted Va values to be
uneven and can potentially go backwards on itself.

Testing this solution by stepping Vs (not connected on a triode)
shows that the Va against Ia plots (independent of Vs) are
reproduced on top of each other. With the previous solution,
the triode plots did not overlay each other as accurately
demonstrating that noise in the "real" Va value impacts the
reproducibility of the plot.

With this change in place, the triode plots of Va against Ia are
smoother and are reproducible.

Signed-off-by: Dean Jenkins <skullandbones99@gmail.com>